+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Tomato Editor Design
 
+## 0.0.12
+
+- SlideCard增加hasMask接口和onClose接口
+- SlideCard和Popver增加zIndex默认值
+- Button增加size: mini
+
+## 0.0.11
+
+- fix SlideCard zIndex问题
+
+## 0.0.10
+
+- 增加类型文件types
+
 ## 0.0.9
 
 - SlideCard组件把triggerDom接口改成rect